Clean up the Morph packaging
