+++ packages/htmlbars-runtime/lib/main.js
@@ -1,5 +1,5 @@
 import { domHelpers } from "./dom_helpers";
-import { Morph } from "./morph";
+import { Morph } from "morph";
 
 export var domHelpers;
 export var Morph;
